(1297) Fix user-access form for territorial coll.
