--- packages/frontend/src/js/app/components/form/input/collectivity/collectivity.pug
@@ -1,1 +1,0 @@
-<Autocompleter v-bind="options" v-model="data" :disabled="disabled"></Autocompleter>
